requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
slack-sdk>=3.21.3
jira>=3.6.1
openai>=1.3.6
